For writes to replicated tables, use primary copy technique to reduce the likelihood of distributed deadlocks. That is, if all writes for a table first go through the same data node, if the same tuple is updated by multiple clients, we can at least ensure that the first session that obtains the lock can similarly obtain the lock for the same tuple on all of the nodes.  (Usual deadlocks are still possible.)
There is a new GUC parameter, primary_data_node. By default it is 1,
the node number where to execute writes to replicated tables first,
before doing the other ones. If it is set to 0, then the primary
node technique is not used, and it will update all replicas
simultaneously.

Instead of the planner returning a list of nodes to execute on,
it returns a pointer to Exec_Nodes, which contains the primary and
secondary nodes to execute on. DataNodeExec() now uses this information.

I also added a new check so that if a different number of rows were
affected on replicated tables (an UPDATE, for example), an error occurs.
This happens for COMBINE_TYPE_SAME. (I tested with the help of EXECUTE
DIRECT, intentionally messing up the data.)