Add support for views.
A view is in effect rewritten into a FROM clause subquery, so we also
add support of detecting safe FROM clause subqueries.

We check after the query tree is rewritten if the query is safe to
execute in a single step, include FROM clause subqueries.
If not, we do not allow it.