Allow rules to be created, provided that they do not use NOTIFY, which is not yet supported.
Note that using rules is a bit unsafe. We currently end up
passing down the original statement (or something close to it)
to the data nodes, but based on the modifications to the
rewritten tree from the rules.

It is possible to do something that violates the distribution rules
of the system. For example, on an update, one could insert into a
table that is hash distributed, but populate it with a value that
violates this.