Support for RENAME/DROP SCHEMA with sequences
Since commit for the support of ALTER SEQUENCE,
sequences use a global name in GTM based on:
db_name.schema_name.sequence_name

This commit permits to rename sequences on GTM
if their schema's name is modified.
This patch permits also to drop a sequence on GTM
in the case that its schema is being dropped in cascade.