Initial support for multi-step queries, including cross-node joins.
Note that this is a "version 1.0" implementation, borrowing some code
from the SQL/MED patch. This means that all cross-node joins
take place on a Coordinator by pulling up data from the data nodes.
Some queries will therefore execute quite slowly, but they will
at least execute.  In this patch, all columns are SELECTed from
the remote table, but at least simple WHERE clauses are pushed
down to the remote nodes. We will optimize query processing
in the future.

Note that the same connections to remote nodes are used in
multiple steps. To get around that problem, we just
add a materialization node above each RemoteQuery node,
and force all results to be fetched first on the
Coordinator.

This patch also allows UNION, EXCEPT and INTERSECT, and other
more complex SELECT statements to run now.

It includes a fix for single-step, multi-node LIMIT and OFFSET.

It also includes EXPLAIN output from the Coordinator's
point of view.

Adding these changes introduced a problem with AVG(),
which is currently not working.