Implementation of 2PC from applications
Support for PREPARE TRANSACTION 'tid',
ROLLBACK PREPARED 'tid'
and COMMIT PREPARED 'tid'.

When a Transaction is prepared on a Coordinator,
the list of involved Datanodes is saved in GTM and transaction is
put in PREPARE state. The transaction ID 'tid' is also saved on GTM.

COMMIT PREPARED or ROLLBACK PREPARED can be issued from a different Coordinator
by using the same tid. The Coordinator receiving the Commit SQL gets a list of Datanodes
from GTM, and commits the transaction on the right nodes.

This patch adds a new interface on GTM to save also the list of Coordinators
involved in a PREPARE transaction.
Coordinator<->Coordinator connection protocol is not implemented yet,
so for the moment Coordinator do not create a 2PC file at PREPARE.
This feature will be added with the implementation of DDL synchronization among Coordinators.