Handle stored functions in queries.
If the stored function is IMMUTABLE and appears in a query,
it can be safely executed on the data nodes and is pushed down.
Otherwise, the stored function must be executed on the coordinator.

Note that stored functions cannot yet contain queries that use
passed in parameters until we add support for prepared statements
with parameters (planned to be done within the next few months).