Special hanlding for ANALYZE on the data nodes.
Before this commit, if autovacuum executed analyze, it would get the
XID from GTM, and the XID would be included in all global snapshots,
even though the XID will never be involved in a transaction on the
other nodes.

This commit adds special handling to track these separately. They
are excluded from global snapshots, and when the data node receives
a snapshot, it adds any of these XIDs to the snapshot.

This helps to prevent any long-running auto-analyze on a single node
slowing down execution on other nodes because of a low xmin value.