This fixes a couple of issues noticed after the last commit, which added checking for expressions in the SELECT clause.
Fxied a bug with expression checking because of context
recursion depth.

Also, we now allow built-in volatile expressions in statements
to be pushed down. Without this, having NOW() in DML was
causing a problem in DBT-1.

We should go back and examine this more carefully to make
sure the built-in ones are all safe.

Also, we still make the false assumption that immutable
functions will not write to the database. This could be addressed
by either a more thorough analysis of functions at
creation time or additional tags at creation time (preferably
in standard PostgreSQL). Ideally such functionality could be
committed back to standard PostgreSQL.