Add support for single-step prepared statements. Works for both named and unnamed prepared statements, works for PREPARE and EXECUTE commands.
The Coordinator tracks a list of the prepared statements,
and prepares them in turn on Data Nodes, and only on demand,
when they are first executed on the target node(s).

At the end of a transaction, if there are still prepared
statements that exist for the session, the connections
are not released to the pool. (We should do something
similar for temporary tables.)

This commit also changes an existing kluge with using the
SQL string in some cases, and now deparses from the Query tree
instead.

Written by Andrei Martsinchyk, multi-step check added by me.