Clean up of execRemote.c
There was some code that was used to clean up connection thread between Coordinator and Datanodes
that was not really necessary.
This has been introduced with version 0.9.2 to stabilize the code by consuming messages on connections
where error happened on backend Noce.

This patch also corrects a bug on Datanode with GXID that was not correctly set
at initialization.
This leaded to transactions being committed twice on backend nodes, crashing it with a FATAL error.

Patch written by Andrei Martsinchyk