Fix for partially committed transactions in Implicit 2PC
When a Datanode crashed during an implicit 2PC,
the list of nodes that crashed were not saved on GTM when aborting
transaction on remote Coordinator.

This problem was introduced by commit 981e2bbe81c94c0427ed9504d0390119a7770a83
in execRemote.c when handle error is set at DN_CONNECTION_STATE_ERROR_FATAL
when a connection to backend was cut.