Fix for bug 3172438: node registration protocol
When a node using the same ID as a node, already running and registered,
tried to register itself GTM returned an error and the node that failed
registration tries to register once again. on GTM.

Between the first and second registration, connection to GTM was reinitialiazed,
making the node running marked as disconnected.

Then the node that failed registration tries a second time and modifies the information
of the node already running.
At this point, GTM had to return an error but the registration succeeds at the second time.

This patch adds some additional control when disconnecting a registered node on GTM
by checking the socket ID of the connection used for node registration.

This permits to keep consistent data in GTM and proxies for nodes in the cluster.