Clean up of pgxc_ddl
This script was used up to version 0.9.2 to make a cold synchronization
of Coordinator catalogs when running a DDL.
At this point, it was necessary to stop the cluster each time a DDL was
launched, resulting in an unefficient process.

This commit removes the dependencies between pgxc_ddl and XC (initdb).
It is now not installed by default and moved to a new repository
called src/pgxc/bin/pgxc_ddl.
A README is also added.