Block DEFERRED constraints (DEFERRABLE)
DEFERRED constraints are checked at transaction COMMIT.
In a Postgres-XC transaction that involved DML failing on
DEFERRED constraints, Datanodes return to Coordinator transaction errors,
this was resulting in transaction partially committed in the cluster.
In the case of pg_regress, test case constraints was putting the cluster in
a stall state.

IMMEDIATE constraints (default) are still available.