Fix for cache lookup bug for type
With this fix, backend Datanode send back to remote Coordinator
the type name and note the OID type to avoid type inconsistencies between
nodes as Postgres-XC is a shared-nothing architecture.

Patch idea is from Mason, base patch is from Benny,
I just gathered all the ideas and improved the patch.