Fix for bug 3234421: multiple user support for pooler
This commit extends pooler functionalities for multiple users.
Now user name used for connection string is built directly
from the user name got on port of postmaster at a user connection on
Coordinator.

This also solves a serious security issue with passwords parameters
in postgresql.conf and when building the connection string of pooler.
It also avoids that passwords are written by pooler logging system.

In consequence, the following GUC parameters are deleted:
data_node_users, data_node_passwords, coordinator_users,
coordinator_passwords.

This commit is a first step for the support of session parameters.