This patch fixes bug ID 3136257.
The test case was trying to prepare a transaction after a fetch from a cursor inside a transaction.

The way cursors are handled in XC, a FETCH can leave one of the datanodes in DN_CONNECTION_STATE_QUERY.
This fact was being ignored by pgxc_all_handles_send_query, which would error out if any of the data nodes
is in any state other than DN_CONNECTION_STATE_IDLE, where as the correct handling in this case is
to call BufferConnection and continue.

Next there was an error in ExecEndRemoteQuery.

PGXCNodePrepare called in response to PREPARE TRANSACTION would call release_handles
to release all data node connections and coordinator connections back to pool.
Later when ExecEndRemoteQuery would try to close cursors, it would find a -1 in sock
of all data node connections and hence the system would report that it was unable to close cursors.

This patch adds correct handling while closing cursors in ExecEndRemoteQuery.