Fix for bug 3269392: GRANT for VIEW and SEQUENCE
GRANT was launched on all the nodes,
but for views and sequence objects it is only necessary to launch it
on Coordinators.

When multiple objects are granted at the same time, XC checks their types
and returns an error in case Coordinator-only objects are mixed with normal
relations.

There is also a fix for regression test privileges.