Support for session and local parameters
This commit adds support for commands like:
SET ROLE ... ;
SET param TO value;
SET SESSION param TO value;
SET LOCAL param TO value;

When a SET command is launched, it is saved in pooler and
then launched to nodes by pooler if connections to backend nodes
exist.
Commands are saved with the following format as a string:
"SET param1 TO value1;...;SET paramN TO valueN"
Local and session commands are saved as separate strings.

When a new connection is created to a backend node, pooler replays
all the saved SET commands.

When a transaction is finished, local parameters are deleted from
pooler. It is not necessary in this case to reset on backend nodes
as transaction commit has made the work.
If a SET command has been launched for a non-local parameter,
connections to nodes are kept alive with the session and not sent back to pool
when a transaction finishes.

When session is finished (user logging off), pooler sends asynchronously
a "RESET ALL" command to each connection and put connections back to pool.
Reset is not launched if no SET queries for session parameters have been launched.

A SET command for local parameters is not sent to pooler if it is not
inside a transaction block to save ressources in the cluster.

This commit contains also a couple of corrections for regression tests
according to implementation of session parameters.