Block use of non-immutable functions as DEFAULT values
DEFAULT is used by CREATE/ALTER TABLE/DOMAIN.

A DML query using a default with non-immutable function needs to be fed by Coordinator.
This is particularly necessary in the case of replicated tables.
But even in the case of distributed tables timestamp values have to be taken from
GTM to avoid data inconsistencies through the cluster.

As a possible implementation, a constraint with non-immutable functions
is just created on Coordinator and not on Datanodes.
When a DML query needs a default value Coordinator feeds it,
rewrites the query, and distributes it to Datanodes
so as the uniqueness of the value is insured.

Sequence (currval, nextval) and timestamp values (now()...) have
to be taken from GTM.