This patch fixes bug ID 3280980 in which ORDER BY BIT fails
The reason of the failure was in the function create_tuple_desc.

The function do_query sends the query to appropriate nodes
and then reads the results. It then calls handle_response
function to process messages received from the nodes.

On receiving the RowDescription message, the function calls
HandleRowDescription to create tuple descriptor from
row description message.

In the function create_tuple_desc type mode was being ignored
as received in the row description.
Instead the function was asking parseTypeString to decode
type modifier from type name.

Although the BIT array was defined to be 11 bits long, the
function would report that type modifier is 1.

This was causing the query to fail.

This patch changes the code so that it now no longer ignores
the type mode received in the row description message.