This patch solves bug ID 3237805
Problem Description:

The function deparse_context_for_remotequery was being used to create
a deparese_context. While creating the rte for deparse_context,
system was passing relation name get_rel_name(rte->relid)
to the function. The function would then use makeAlias
and was using it in eref of the rte.

This would pass table name in eref although user had specified
an alias. This meant that the query
   SELECT p.hobbies FROM person p;
would generate this query for data nodes
   SELECT person.* FROM public.person p;
which is an error
ERROR:  invalid reference to FROM-clause entry for table "person"

The solution was to pass rte->eref to the function.