Extension of CLEAN CONNECTION for specific users
New synopsis of	this query becomes:
CLEAN CONNECTION TO
   (COORDINATOR num | DATANODE num | ALL {FORCE})
   [ FOR DATABASE dbname ]
   [ TO USER username ];

This increases cleaning granularity of XC cluster.
This query cannot be launched if no database or no user are specified.

If only a user name is specified, cleaning is made for all the databases
of this user.
If only a database name is specified, cleaning is made for all the users
of this database.
It is also possible to specify both database and user name.

This commit also contains modifications for cleaning connections
when dropping a database.
Now before dropping a database, CLEAN CONNECTION is launched
first on all the Coordinators.
Then drop query is sent to backend nodes.