Fix for bug 3231445: COPY TO error when combining data
The origin of this bug was round robin table.
In the case of COPY TO, connections to all the datanodes
of the table are necessary but it happened that only one node
was picked up.

COPY can be applied on a single node only for COPY TO under
a replicated table.

Patch written by rahuahua