Fix for bug 3167720: join optimizer error
When running a query on multiple tables, if a constant
expression was contained in WHERE clause, XC planner
examined the constant expression all the time and checked
if it was possible to avoid multiple node operation
if the constant expression depended on a distributed table.

This commits adds a check to verify in the case of multiple tables
used in FROM clause if query can be safely pushed down to one
target node on depending on the distribution type of each table.

If SELECT clause is used on one distributed table and several
replicated tables, this can be safely pushed down to a target node
but in the case of multiple distributed tables used, all the nodes
have to be targetted.