Fix for bug 3289281: DROP TABLE in prepared transactions
In a cluster configuration using a single coordinator,
it happened that a transaction involving DDL was not prepared
on the single Coordinator but directly committed.

This patch fixes this issue by also taking into account special handling
such as transactions using only view or sequence DDLs.

This commit contains at the same time a fix for bug 3273585.