This patch adds support for the following data types to be used as distribution key
INT8, INT2, OID, INT4, BOOL, INT2VECTOR, OIDVECTOR
CHAR, NAME, TEXT, BPCHAR, BYTEA, VARCHAR
FLOAT4, FLOAT8, NUMERIC, CASH
ABSTIME, RELTIME, DATE, TIME, TIMESTAMP, TIMESTAMPTZ, INTERVAL, TIMETZ

A new function compute_hash is added in the system which is used to
compute hash of a any of the supported data types.
The computed hash is used in the function GetRelationNodes to
find the targeted data node.

EXPLAIN for RemoteQuery has been modified to show the number of
data nodes targeted for a certain query. This is essential
to spot bugs in the optimizer in case it is targeting all nodes
by mistake.

In case of optimisations where comparison with a constant leads
the optimiser to point to a single data node, there were a couple
of mistakes in examine_conditions_walker.
First it was not supporting RelabelType, which represents a "dummy"
type coercion between two binary compatible datatypes.
This was resulting in the optimization not working for varchar
type for example.
Secondly it was not catering for the case where the user specifies the
condition such that the constant expression is written towards LHS and the
variable towards the RHS of the = operator.
i.e. 23 = a

A number of test cases have been added in regression to make sure
further enhancements do not break this functionality.

This change has a sizeable impact on current regression tests in the following manner.

1. horology test case crashes the server and has been commented out in serial_schedule.
2. In money test case the planner optimizer wrongly kicks in to optimize this query
   SELECT m = '$123.01' FROM money_data;
   to point to a single data node.
3. There were a few un-necessary EXPLAINs in create_index test case.
   Since we have added support in EXPLAIN to show the number of
   data nodes targeted for RemoteQuery, this test case was producing
   output dependent on the cluster configuration.
4. In guc test case
   DROP ROLE temp_reset_user;
   results in
   ERROR:  permission denied to drop role